Code reorg done
